Module Rsa.Private

The format for private keys is:

4 bytes: size of d encoded in big endian

d

3 bytes: e (0x01 0x00 0x01)

p

q

d, p and q are encoded with a leading 0x00. The size of p and q is determined from that of d (|p| = |q| = |d|/2 + 1).

The format is a bit ambiguous if e is not 0x010001, so an error will be raised in that case.

type t = {

e : Z.t;

d : Z.t;

p : Z.t;

q : Z.t;

}

val equal : t -> t -> Ppx_deriving_runtime.bool

val compare : t -> t -> Ppx_deriving_runtime.int

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val decode : Cstruct.t -> ( t, string ) Result.result